Moved abs and app into 'named' sections, in preparation for debruijn
