+++ example-strict-lc/eval/Lambda/Eval.hs
@@ -3,8 +3,8 @@ module Lambda.Eval where
 import Control.Applicative
 import Control.MaybeK
 import Lambda.Term
-import Lambda.Abs.Semantics
-import Lambda.App.Semantics.Strict
+import Lambda.Abs.Named.Semantics
+import Lambda.App.Named.Semantics.Strict
 import Semantics.BigStep
 
 eval :: Term -> Term
