+++ lambda-abs-sig/Lambda/Abs/Named/Semantics.hs
@@ -1,8 +1,8 @@
-module Lambda.Abs.Semantics where
+module Lambda.Abs.Named.Semantics where
 
 import Control.MaybeK
 import Lambda.Term
-import Lambda.Abs
+import Lambda.Abs.Named
 import Semantics.BigStep
 
 absBigStep :: (TermF f -> MaybeK (TermF f)) -> BigStep (TermF f) (TermF f)
