+++ lambda-app-sig/Lambda/App/Named/Semantics/Lazy.hs
@@ -1,11 +1,11 @@
-module Lambda.App.Semantics.Lazy where
+module Lambda.App.Named.Semantics.Lazy where
 
 import Control.Applicative
 import Control.Lens
 import Control.MaybeK
 import Lambda.Term
 import Lambda.App
-import Lambda.Abs
+import Lambda.Abs.Named
 import Lambda.Subst
 import Semantics.BigStep
